+++ gno.land/pkg/integration/utils.go
@@ -1,21 +1,10 @@
 package integration
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"strconv"
 	"strings"
-	"testing"
-
-	"github.com/gnolang/gno/tm2/pkg/crypto"
-	"github.com/gnolang/gno/tm2/pkg/crypto/bip39"
-	"github.com/gnolang/gno/tm2/pkg/crypto/hd"
-	"github.com/gnolang/gno/tm2/pkg/crypto/secp256k1"
-	"github.com/stretchr/testify/require"
 )
 
 // `unquote` takes a slice of strings, resulting from splitting a string block by spaces, and
@@ -81,65 +70,4 @@ func unquote(args []string) ([]string, error) {
 	}
 
 	return parts, nil
-}
-
-func buildGnoland(t *testing.T, rootdir string) string {
-	t.Helper()
-
-	bin := filepath.Join(t.TempDir(), "gnoland-test")
-
-	t.Log("building gnoland integration binary...")
-
-	// Build a fresh gno binary in a temp directory
-	gnoArgsBuilder := []string{"build", "-o", bin}
-
-	os.Executable()
-
-	// Forward `-covermode` settings if set
-	if coverMode := testing.CoverMode(); coverMode != "" {
-		gnoArgsBuilder = append(gnoArgsBuilder,
-			"-covermode", coverMode,
-		)
-	}
-
-	// Append the path to the gno command source
-	gnoArgsBuilder = append(gnoArgsBuilder, filepath.Join(rootdir,
-		"gno.land", "pkg", "integration", "process"))
-
-	t.Logf("build command: %s", strings.Join(gnoArgsBuilder, " "))
-
-	cmd := exec.Command("go", gnoArgsBuilder...)
-
-	var buff bytes.Buffer
-	cmd.Stderr, cmd.Stdout = &buff, &buff
-	defer buff.Reset()
-
-	if err := cmd.Run(); err != nil {
-		require.FailNowf(t, "unable to build binary", "%q\n%s",
-			err.Error(), buff.String())
-	}
-
-	return bin
-}
-
-// GeneratePrivKeyFromMnemonic generates a crypto.PrivKey from a mnemonic.
-func GeneratePrivKeyFromMnemonic(mnemonic, bip39Passphrase string, account, index uint32) (crypto.PrivKey, error) {
-	// Generate Seed from Mnemonic
-	seed, err := bip39.NewSeedWithErrorChecking(mnemonic, bip39Passphrase)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate seed: %w", err)
-	}
-
-	// Derive Private Key
-	coinType := crypto.CoinType // ensure this is set correctly in your context
-	hdPath := hd.NewFundraiserParams(account, coinType, index)
-	masterPriv, ch := hd.ComputeMastersFromSeed(seed)
-	derivedPriv, err := hd.DerivePrivateKeyForPath(masterPriv, ch, hdPath.String())
-	if err != nil {
-		return nil, fmt.Errorf("failed to derive private key: %w", err)
-	}
-
-	// Convert to secp256k1 private key
-	privKey := secp256k1.PrivKeySecp256k1(derivedPriv)
-	return privKey, nil
 }
